Add omit_term_stderr config (#79)
To disable unnecessary backend terminal output to the stderr.

There are some annoying but harmless logs from kitty like:

"""
[0.201] The output buffer does not support sRGB color encoding, colors will be incorrect.
[0.321] [PARSE ERROR] Escape codes to resize text area are not supported
"""

which are printed all the time. See:
https://github.com/kovidgoyal/kitty/issues/7409
https://github.com/kovidgoyal/kitty/issues/8545

Just give user an option to disable the stderr output. The error log
will be disabled if the option is true and the log level is above
DEBUG.

(Have thought about adding a log filter to specific backend, but it
seems to be over-designed.)